Move some code to their respective file
